requirements.txt
pyserial
matplotlib
python3-tk
pandas